+++ crates/builder/core/src/flashblocks/handler.rs
@@ -8,7 +8,8 @@ use tracing::warn;
 ///
 /// In the case of a payload built by this node, it is broadcast to peers and an event is sent to the payload builder.
 /// In the case of a payload received from a peer, it is executed and if successful, an event is sent to the payload builder.
-pub(super) struct PayloadHandler {
+#[derive(Debug)]
+pub struct PayloadHandler {
     // receives new payloads built by this builder.
     built_rx: mpsc::Receiver<OpBuiltPayload>,
     // sends a `Events::BuiltPayload` to the reth payload builder when a new payload is received.
@@ -16,14 +17,17 @@ pub(super) struct PayloadHandler {
 }
 
 impl PayloadHandler {
-    pub(super) const fn new(
+    /// Constructs a new `PayloadHandler`.
+    pub const fn new(
         built_rx: mpsc::Receiver<OpBuiltPayload>,
         payload_events_handle: tokio::sync::broadcast::Sender<Events<OpEngineTypes>>,
     ) -> Self {
         Self { built_rx, payload_events_handle }
     }
 
-    pub(crate) async fn run(self) {
+    /// Runs the payload handler, listening for new built payloads
+    /// and forwarding them to the payload builder via events.
+    pub async fn run(self) {
         let Self { mut built_rx, payload_events_handle } = self;
 
         tracing::debug!("flashblocks payload handler started");
